Tests: move to folder, split out validation tests
Also added a new test for rejected responses when a provider throws an error
